pyproject.toml
[project]
name = "md-reports"
version = "0.1.0"
description = "Render Markdown reports to multiple formats (DOCX, with pluggable renderers)"
readme = "README.md"
authors = [
    { name = "pythoro", email = "pythoro@mindquip.com" }
]
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1",
    "markdown-it-py>=3.0",
    "mdit-py-plugins>=0.4",
    "python-docx>=1.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "ruff>=0.5",
]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[tool.ruff]
line-length = 79
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
